requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
cbor==1.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cmake==4.0.2
colorama==0.4.6
datasets==3.6.0
dill==0.3.8
diskcache==5.6.3
duckdb==1.3.0
et_xmlfile==2.0.0
faiss-cpu==1.7.4
fastapi==0.115.12
filelock==3.18.0
FlagEmbedding==1.3.5
frozenlist==1.6.2
fsspec==2025.3.0
grpcio==1.72.1
h11==0.16.0
h2==4.2.0
hf_transfer==0.1.9
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.4
hyperframe==6.1.0
idna==3.10
ijson==3.4.0
inscriptis==2.6.0
ir_datasets==0.5.10
Jinja2==3.1.6
joblib==1.5.1
llama_cpp_python @ file:
lxml==5.4.0
lz4==4.4.4
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.5
ninja==1.11.1.4
numpy==1.26.4
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
peft==0.15.2
pillow==11.2.1
portalocker==2.10.1
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.52.4
trec-car-tools==2.6
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
unlzw3==0.2.3
urllib3==2.4.0
uvicorn==0.34.3
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchfiles==1.0.5
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
zlib-state==0.1.9
